Добавляет smart компонент
